obs space to float32 issue #24 (#25)

+++ craftax/craftax/envs/craftax_pixels_env.py
@@ -80,7 +80,7 @@ class CraftaxPixelsEnvNoAutoReset(EnvironmentNoAutoReset):
                 (OBS_DIM[0] + INVENTORY_OBS_HEIGHT) * BLOCK_PIXEL_SIZE_AGENT,
                 3,
             ),
-            dtype=jnp.int32,
+            dtype=jnp.float32,
         )
 
 
@@ -152,5 +152,5 @@ class CraftaxPixelsEnv(environment.Environment):
                 (OBS_DIM[0] + INVENTORY_OBS_HEIGHT) * BLOCK_PIXEL_SIZE_AGENT,
                 3,
             ),
-            dtype=jnp.int32,
+            dtype=jnp.float32,
         )
