+++ craftax/craftax/envs/craftax_symbolic_env.py
@@ -104,7 +104,7 @@ class CraftaxSymbolicEnvNoAutoReset(EnvironmentNoAutoReset):
             0.0,
             1.0,
             (obs_shape,),
-            dtype=jnp.int32,
+            dtype=jnp.float32,
         )
 
 
@@ -177,5 +177,5 @@ class CraftaxSymbolicEnv(environment.Environment):
             0.0,
             1.0,
             (obs_shape,),
-            dtype=jnp.int32,
+            dtype=jnp.float32,
         )
